+++ src/roq/flags.h
@@ -1,0 +1,11 @@
+/* Copyright (c) 2017-2021, Hans Erik Thrane */
+
+#pragma once
+
+#include "roq/flags/flags.h"
+
+namespace roq {
+
+using flags::Flags;
+
+}  // namespace roq