+++ src/roq/tool.cpp
@@ -33,8 +33,10 @@ Tool::Tool(
     : args_(initialize_flags(argc, argv, description, version)), build_type_(build_type), git_hash_(git_hash),
       compile_date_(compile_date), compile_time_(compile_time) {
   assert(args_.size() > 0);
-  auto pattern = "%v"_sv;
-  Logger::initialize(args_[0], pattern);
+  Logger::Config config{
+      .pattern = "%v"_sv,
+  };
+  Logger::initialize(args_[0], config);
 }
 
 Tool::~Tool() {